requirements.txt
Flask
python-dotenv
redis
gunicorn
prometheus-client